Move uefi_common to library.uefi
