+++ chipsec/library/uefi/uefi_common.py
