+++ chipsec/utilcmd/uefi_cmd.py
@@ -59,7 +59,7 @@ import uuid
 from argparse import ArgumentParser
 
 from chipsec.command import BaseCommand, toLoad
-from chipsec.hal.common.uefi_common import EFI_STATUS_DICT, parse_efivar_file
+from chipsec.library.uefi.uefi_common import EFI_STATUS_DICT, parse_efivar_file
 from chipsec.library.file import write_file, read_file
 from chipsec.hal.common.spi_uefi import decode_uefi_region, modify_uefi_region, compress_image, CMD_UEFI_FILE_REPLACE
 from chipsec.hal.common.spi_uefi import CMD_UEFI_FILE_INSERT_AFTER, CMD_UEFI_FILE_INSERT_BEFORE, CMD_UEFI_FILE_REMOVE
